requirements.txt
flask==2.3.2
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.1.0
bleach==6.1.0
python-magic==0.4.27
flask-limiter==3.5.0
colorlog==6.8.0
cachetools==5.3.2